python/notebooks/marimo/requirements.txt
beautifulsoup4
click
decouple
marimo
matplotlib
numpy
pandas
psycopg2-binary
pydantic
python-dotenv
requests
scikit-learn
sqlmodel
types-psycopg2